feat : supabase client setup
